# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My DOC API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/": {},
        "/api/auth/register": {
            "post": {
                "tags": [
                    "API ALL"
                ],
                "summary": "Регистрация",
                "operationId": "b6d359549f01fa78e74b5d268699917e",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "name"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "email@email.com"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "password"
                                            },
                                            "password_confirmation": {
                                                "type": "string",
                                                "example": "password"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User successfully registered"
                                        },
                                        "user": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "name"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "email@email.com"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2024-03-31T18:18:06.000000Z"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-03-31T18:18:06.000000Z"
                                                },
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "API ALL"
                ],
                "summary": "Авторизация",
                "operationId": "4ed2d4ad482dbe329dc4b3dd1e97ba7c",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "email": {
                                                "type": "string",
                                                "example": "test@test.test"
                                            },
                                            "password": {
                                                "type": "string",
                                                "example": "testtest"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "{{ token }}"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": 3600
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "name"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "email@email.com"
                                                },
                                                "email_verified_at": {
                                                    "type": "string",
                                                    "example": "2024-03-31T18:18:06.000000Z"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-03-31T18:18:06.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2024-03-31T18:18:06.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/refresh": {
            "post": {
                "tags": [
                    "API ALL"
                ],
                "summary": "Обновление токена",
                "operationId": "e9c0fcf285155c43948b0e65fa077c5e",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "{{ token }}"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": 3600
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "name"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "email@email.com"
                                                },
                                                "email_verified_at": {
                                                    "type": "string",
                                                    "example": "2024-03-31T18:18:06.000000Z"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-03-31T18:18:06.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2024-03-31T18:18:06.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/user-profile": {
            "get": {
                "tags": [
                    "API ALL"
                ],
                "summary": "Информация о пользователе",
                "operationId": "336893717444bc0220c2b322a74777ae",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "name"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "email@email.com"
                                        },
                                        "email_verified_at": {
                                            "type": "string",
                                            "example": "2024-03-31T18:18:06.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2024-03-31T18:18:06.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2024-03-31T18:18:06.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "API ALL"
                ],
                "summary": "Выход",
                "operationId": "76091f15f3ecbb84edf45f4bb0cfe8ce",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User successfully signed out"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}